+++ src/shared/components/Pixedi/hooks/index.ts
@@ -2,3 +2,5 @@ export { useImageLoader } from "./useImageLoader";
 export { useMobile } from "./useMobile";
 export { useBellow } from "./useBellow";
 export { useAbove } from "./useAbove";
+export { useImageSaving } from "./useImageSaving";
+export { useToolCommit } from "./useToolCommit";