+++ src/shared/components/Pixedi/provider/initialState.ts
@@ -1,3 +1,4 @@
+import { DefaultFilterValues } from "../filters";
 import { type History, type Action, ActionName } from "../types";
 
 import type { Settings } from "../types";
@@ -12,6 +13,47 @@ export type PixediContextType = {
   settings: Settings;
   isAlpha: boolean;
   showCompare: boolean;
+  singleToolUI: boolean;
+};
+
+const getInitialAction = (
+  tools: string[],
+  width: number,
+  height: number,
+): Action | null => {
+  if (tools.length === 1) {
+    switch (tools[0]) {
+      case "resize":
+        return { name: ActionName.RESIZE, args: { width, height } };
+      case "crop":
+        return {
+          name: ActionName.CROP,
+          args: { id: "freeform", ratio: width / height, isFree: true },
+        };
+      case "presetCrop":
+        return {
+          name: ActionName.PRESET_CROP,
+          args: {
+            id: "facebook-post",
+            ratio: 1200 / 630,
+            isFree: false,
+            preset: { width: 1200, height: 630 },
+          },
+        };
+      case "flip":
+        return {
+          name: ActionName.FLIP,
+          args: { horizontal: false, vertical: false },
+        };
+      case "rotate":
+        return { name: ActionName.ROTATE, args: { degrees: 0 } };
+      case "filters":
+        return { name: ActionName.FILTERS, args: DefaultFilterValues() };
+      default:
+        return null;
+    }
+  }
+  return null;
 };
 
 export const initialSettings: Settings = {
@@ -44,7 +86,7 @@ export const getInitialState = (
       },
     ],
   },
-  currentAction: null,
+  currentAction: getInitialAction(settings?.tools || [], width, height),
   previewUrl,
   originalBlob,
   mimeType,
@@ -52,4 +94,5 @@ export const getInitialState = (
   settings,
   isAlpha,
   showCompare: false,
+  singleToolUI: settings?.tools?.length === 1,
 });